services/user-service/requirements.txt
# FastAPI and server
fastapi[standard]==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# Email validation
email-validator==2.2.0

# Testing (optional, but recommended)
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2

